% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cases.R
\name{assign_cases}
\alias{assign_cases}
\title{Probabilistically assign rows in a dataset to a "target" geography}
\usage{
assign_cases(source, source_id, xwalk_df)
}
\arguments{
\item{source}{data.frame containing rows to transfer to \code{target}}

\item{source_id}{character. Column of ids that link source to xwalk_df}

\item{xwalk_df}{data.frame. Must have the following columns: source_id,
target_id, s2t_fraction. Use create_xwalk to generate the crosswalk.}
}
\value{
a vector of values representing the \code{target_id} to assign the row
}
\description{
Probabilistically assign rows in a dataset to a "target" geography
}
\details{
This function uses the s2t_fraction to probabilistically assign rows to a target id

set the seed \code{set.seed} to make things deterministically random.
}
