% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bbox_to_sf}
\alias{bbox_to_sf}
\title{Convert a bounding box to an sf data frame}
\usage{
bbox_to_sf(bbox, id = 1)
}
\arguments{
\item{bbox}{a result from st_bbox}
}
\description{
Convert a bounding box to an sf data frame
}
