% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_internal_consistency}
\alias{check_internal_consistency}
\title{check the internal consitency of an sf file}
\usage{
check_internal_consistency(x, stop = TRUE)
}
\arguments{
\item{x}{a sf data.frame}

\item{stop}{logical. Should this function throw an error or return the results?}
}
\description{
check the internal consitency of an sf file
}
