% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk.R
\name{create_xwalk}
\alias{create_xwalk}
\title{Create an xwalk weight matrix to crosswalk estimates between non-nesting geographies}
\usage{
create_xwalk(
  source,
  target,
  source_id = "GEOID",
  target_id = "GEOID",
  method = "fractional overlap",
  point_pop = if (requireNamespace("kcparcelpop")) kcparcelpop::parcel_pop else NULL,
  min_overlap = 0.75,
  ...
)
}
\arguments{
\item{source}{sf data.frame. Sf object containing the source geographies}

\item{target}{sf data.frame. sf object containng the target geographies}

\item{source_id}{character. Column that has the unique ids in source}

\item{target_id}{character. Column that has the unique ids in target}

\item{method}{character. Method for generating the crosswalk estimate. One of "point pop" and "fractional overlap"}

\item{point_pop}{sf data.frame with a "pop" column representing the number of people living at/around that point}

\item{min_overlap}{numeric between 0 and 1. Percent of the bounding boxes of source and target that need to overlap. Otherwise, throw an error}

\item{...}{additional arguments passed to underlying methods. xwalk_polap or xwalk_folap}
}
\description{
Create an xwalk weight matrix to crosswalk estimates between non-nesting geographies
}
