% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk estimates between non-nesting geographies}
\usage{
crosswalk(
  source,
  source_id,
  est,
  se = NULL,
  by = NULL,
  xwalk_df,
  rescale = TRUE
)
}
\arguments{
\item{source}{data.frame containing estimates to transfer to \code{source}}

\item{source_id}{character. Column of ids that link source to xwalk_df}

\item{est}{character. Column in \code{source} containing the estimates for transfer}

\item{se}{character \link{optional}. The standard error of the estimate in \code{source}}

\item{by}{character \link{optional}. Vector of column names to compute by. A given id must be unique within the combination of by.}

\item{xwalk_df}{data.frame. Must have at least three columns: source_id, target_id, and s2t_fraction. Use create_xwalk to generate the crosswalk.}

\item{rescale}{logical. Should the crosswalk weights (s2t_fraction) by scaled to 1 per target? Weights that don't initially sum to 1 are usually when source doesn't totally cover target}
}
\description{
Crosswalk estimates between non-nesting geographies
}
\details{
By computations happen all at once-- that is, if c('race','sex') is passed, the computations are by the intersection of race and sex, not as race and then sex.

Rescale is generally advised when "est" is not a count. When 'est' is a count, mileage may vary.
}
