% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reduce_overlaps}
\alias{reduce_overlaps}
\title{Try to eliminate internal/self overlaps}
\usage{
reduce_overlaps(x, snap = 0.5)
}
\arguments{
\item{x}{sf object}

\item{snap}{distince in st_crs(x,parameters = T)$ud_unit units to allow for snapping}
}
\value{
An sf object that hopefully passes check_internal_consistency checks.
}
\description{
Try to eliminate internal/self overlaps
}
\details{
Users should visually review the resulting object to make sure the hueristics used here weren't too goofy
}
