% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatagg-package.R
\docType{package}
\name{spatagg-package}
\alias{spatagg}
\alias{spatagg-package}
\title{spatagg: Aggregation between non-nesting geographies}
\description{
Provides routines for the aggregation of estimates between sets of non-nesting geographies.
}
\author{
\strong{Maintainer}: Daniel Casey \email{dcasey@kingcounty.gov}

}
\keyword{internal}
