% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_col}
\alias{validate_col}
\title{A function to validate a column exists in a dataset and is unique}
\usage{
validate_col(d, id, type = "source")
}
\description{
A function to validate a column exists in a dataset and is unique
}
