% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk.R
\name{xwalk_polap}
\alias{xwalk_polap}
\title{Create an xwalk matrix based on population overlap}
\usage{
xwalk_polap(
  source,
  target,
  source_id = "id",
  target_id = "id",
  point_pop,
  pp_min_overlap = 0.75,
  ...
)
}
\arguments{
\item{source}{sf data.frame. Sf object containing the source geographies}

\item{target}{sf data.frame. sf object containng the target geographies}

\item{source_id}{character. Column that has the unique ids in source}

\item{target_id}{character. Column that has the unique ids in target}

\item{point_pop}{sf data.frame with a "pop" column representing the number of people living at/around that point}

\item{pp_min_overlap}{numeric between 0 and 1. Percent of the bounding boxes of source and target that need to overlap with the bbox from point_pp. Otherwise, throw an error}
}
\value{
data.frame with columns from target, source, and a \code{fraction}
denoting the fraction of the source population that lives in source.
Depending on how the two geography sets overlap, it may be that source doesn't
fully cover target
}
\description{
Create an xwalk matrix based on population overlap
}
