% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk estimates between non-nesting geographies}
\usage{
crosswalk(
  source,
  source_id,
  est,
  proportion = FALSE,
  se = NULL,
  by = NULL,
  xwalk_df,
  rescale = TRUE
)
}
\arguments{
\item{source}{data.frame containing estimates to transfer to \code{target}}

\item{source_id}{character. Column of ids that link source to xwalk_df}

\item{est}{character. Column in \code{source} containing the estimates for transfer}

\item{proportion}{logical. Does \code{est} represent proportion (0 - 1) data?}

\item{se}{character (optional). The standard error of the estimate in \code{source}}

\item{by}{character (optional). Vector of column names to compute by. A given id must be unique within the combination of by.}

\item{xwalk_df}{data.frame. Must have the following columns: source_id, target_id, s2t_fraction, isect_amount, tcoverage_amount, target_amount. Use create_xwalk to generate the crosswalk.}

\item{rescale}{logical. Should the crosswalk weights by scaled to encompass all of target?}
}
\description{
Crosswalk estimates between non-nesting geographies
}
\details{
By computations happen all at once-- that is, if c('race','sex') is passed, the computations are by the intersection of race and sex, not as race and then sex.

When rescale is TRUE, and tcoverage_amount != target_amount for a given source -> target pair
the crosswalk weights will be scaled up by target_amount/tcoverage_amount

Crosswalk weights for \code{proportion = FALSE} are \code{s2t_fraction} optionally scaled via rescale.
For \code{proportion = TRUE}, the crosswalk weights are \code{isect_amount/target_amount}. The rescale option is irrelevant.
}
